+++ src/stores/session.ts
@@ -1,10 +1,16 @@
 import { reactive } from 'vue';
+import { RouteRecordRaw } from 'vue-router';
 
 const state = reactive({
   /**
    * Indicates if the current page has a page config.
    */
   hasPageConfig: false,
+
+  /**
+   * The last 5 opened routes.
+   */
+  lastOpenedRoutes: [] as RouteRecordRaw[],
 });
 
 /**
@@ -16,6 +22,27 @@ export const useVasSessionStore = () => {
 
     setHasPageConfig(value: boolean) {
       state.hasPageConfig = value;
+    },
+
+    addLastOpenedRoute(route: RouteRecordRaw) {
+      if (!route.name || route.meta?.hideInStyleguide) {
+        return;
+      }
+
+      const existingIndex = state.lastOpenedRoutes.findIndex((r) => r.name === route.name);
+
+      if (existingIndex !== -1) {
+        state.lastOpenedRoutes.splice(existingIndex, 1);
+      }
+
+      state.lastOpenedRoutes.unshift({
+        ...route,
+        children: [], // Clear children for the "Last Opened" group
+      });
+
+      if (state.lastOpenedRoutes.length > 5) {
+        state.lastOpenedRoutes.pop();
+      }
     },
   };
 };