+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- [FEATURE] Add "Last 5 Opened" pages group to the sidebar navigation.
 - [ENHANCEMENT] Add the possibility to mark routes as favorite.
 - [ENHANCEMENT] Split type imports and value imports.
 - [ENHANCEMENT] Add the possibility to custom sort styleguide routes in the sidebar.